Fill in impls in gen partially (#11)
* Impl hash_vrf_base

* Impl pub API compute_vrf()

* Fix test as Network is needless when creating new Address

* Impl is_round_winner a bit

* Fill in compute_proof

* Partial is_ticket_winner impl

* .

+++ core/sectorbuilder/src/fs/mod.rs
@@ -179,7 +179,7 @@ impl FS {
         let c_path = CString::new(path.as_os_str().to_str().unwrap()).unwrap();
         let value = unsafe { statfs(c_path.as_ptr(), &mut fs_stat) };
         if value == 0 {
-            let fsavail = fs_stat.f_bavail as i64 * fs_stat.f_bsize;
+            let fsavail = fs_stat.f_bavail as i64 * fs_stat.f_bsize as i64;
             let avail = fsavail - self.reserved_bytes(path);
             Some((avail, fsavail))
         } else {
